Refactor: init shell
